refactor: refactor sdk3rd client implementations to use a unified signer structure
