+++ pkg/sdk3rd/1panel/client.go
@@ -1,9 +1,7 @@
 package onepanel
 
 import (
-	"crypto/md5"
 	"crypto/tls"
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"net/http"
@@ -36,22 +34,23 @@ func NewClient(serverUrl string, optFns ...OptionsFunc) (*Client, error) {
 		return nil, fmt.Errorf("sdkerr: unset apiKey")
 	}
 
-	restyClient := resty.New().
+	signer := &signer{
+		apiKey: opts.ApiKey,
+	}
+	httper := resty.New().
 		SetBaseURL(strings.TrimSuffix(serverUrl, "/")+"/api/v1").
 		SetHeader("Accept", "application/json").
 		SetHeader("Content-Type", "application/json").
 		SetHeader("User-Agent", app.AppUserAgent).
-		SetPreRequestHook(func(c *resty.Client, req *http.Request) error {
-			timestamp := fmt.Sprintf("%d", time.Now().Unix())
-			tokenMd5 := md5.Sum([]byte("1panel" + opts.ApiKey + timestamp))
-			tokenMd5Hex := hex.EncodeToString(tokenMd5[:])
-			req.Header.Set("1Panel-Timestamp", timestamp)
-			req.Header.Set("1Panel-Token", tokenMd5Hex)
+		SetPreRequestHook(func(_ *resty.Client, req *http.Request) error {
+			if err := signer.Sign(req); err != nil {
+				return fmt.Errorf("sdkerr: sign error: %w", err)
+			}
 
 			return nil
 		})
 
-	return &Client{rc: restyClient}, nil
+	return &Client{rc: httper}, nil
 }
 
 func (c *Client) SetTimeout(timeout time.Duration) *Client {
