+++ pkg/sdk3rd/btpanelgo/client.go
@@ -39,7 +39,7 @@ func NewClient(serverUrl string, optFns ...OptionsFunc) (*Client, error) {
 		return nil, fmt.Errorf("sdkerr: unset apiKey")
 	}
 
-	restyClient := resty.New().
+	httper := resty.New().
 		SetBaseURL(strings.TrimSuffix(serverUrl, "/")).
 		SetHeader("Accept", "application/json").
 		SetHeader("Content-Type", "application/x-www-form-urlencoded").
@@ -47,7 +47,7 @@ func NewClient(serverUrl string, optFns ...OptionsFunc) (*Client, error) {
 
 	return &Client{
 		apiKey: opts.ApiKey,
-		rc:     restyClient,
+		rc:     httper,
 	}, nil
 }
 
@@ -97,9 +97,9 @@ func (c *Client) newRequest(method string, path string, params any, multipart bo
 		}
 	}
 
-	timestamp := time.Now().Unix()
-	data["request_time"] = fmt.Sprintf("%d", timestamp)
-	data["request_token"] = generateSignature(fmt.Sprintf("%d", timestamp), c.apiKey)
+	timestamp := fmt.Sprintf("%d", time.Now().Unix())
+	data["request_time"] = timestamp
+	data["request_token"] = generateSignature(timestamp, c.apiKey)
 
 	req := c.rc.R()
 	req.Method = method
