+++ pkg/sdk3rd/lecdn/v3/master/client.go
@@ -53,7 +53,7 @@ func NewClient(serverUrl string, optFns ...OptionsFunc) (*Client, error) {
 		SetHeader("Accept", "application/json").
 		SetHeader("Content-Type", "application/json").
 		SetHeader("User-Agent", app.AppUserAgent).
-		SetPreRequestHook(func(c *resty.Client, req *http.Request) error {
+		SetPreRequestHook(func(_ *resty.Client, req *http.Request) error {
 			if client.accessToken != "" {
 				req.Header.Set("Authorization", "Bearer "+client.accessToken)
 			}
