+++ pkg/sdk3rd/upyun/console/client.go
@@ -44,7 +44,7 @@ func NewClient(optFns ...OptionsFunc) (*Client, error) {
 		SetHeader("Accept", "application/json").
 		SetHeader("Content-Type", "application/json").
 		SetHeader("User-Agent", app.AppUserAgent).
-		SetPreRequestHook(func(c *resty.Client, req *http.Request) error {
+		SetPreRequestHook(func(_ *resty.Client, req *http.Request) error {
 			if client.loginCookie != "" {
 				req.Header.Set("Cookie", client.loginCookie)
 			}
