+++ pkg/sdk3rd/wangsu/openapi/client.go
@@ -1,17 +1,9 @@
 package openapi
 
 import (
-	"bytes"
-	"crypto/hmac"
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"io"
 	"net/http"
-	"net/url"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/go-resty/resty/v2"
@@ -36,85 +28,24 @@ func NewClient(optFns ...OptionsFunc) (*Client, error) {
 		return nil, fmt.Errorf("sdkerr: unset secretKey")
 	}
 
-	restyClient := resty.New().
+	signer := &signer{
+		accessKey: options.AccessKey,
+		secretKey: options.SecretKey,
+	}
+	httper := resty.New().
 		SetBaseURL("https://open.chinanetcenter.com").
 		SetHeader("Accept", "application/json").
 		SetHeader("Content-Type", "application/json").
 		SetHeader("User-Agent", app.AppUserAgent).
-		SetPreRequestHook(func(c *resty.Client, req *http.Request) error {
-			// API 签名机制：
-			// https://www.wangsu.com/document/openapi/api-authentication
-
-			method := strings.ToUpper(req.Method)
-
-			path := "/"
-			if req.URL != nil {
-				path = req.URL.Path
+		SetPreRequestHook(func(_ *resty.Client, req *http.Request) error {
+			if err := signer.Sign(req); err != nil {
+				return fmt.Errorf("sdkerr: sign error: %w", err)
 			}
-
-			queryStr := ""
-			if method != http.MethodPost && req.URL != nil {
-				queryStr = req.URL.RawQuery
-
-				s, err := url.QueryUnescape(queryStr)
-				if err != nil {
-					return err
-				}
-
-				queryStr = s
-			}
-
-			canonicalHeaders := "" +
-				"content-type:" + strings.TrimSpace(strings.ToLower(req.Header.Get("Content-Type"))) + "\n" +
-				"host:" + strings.TrimSpace(strings.ToLower(req.Host)) + "\n"
-			signedHeaders := "content-type;host"
-
-			payloadStr := ""
-			if method != http.MethodGet && req.Body != nil {
-				payloadb, err := io.ReadAll(req.Body)
-				if err != nil {
-					return err
-				}
-
-				payloadStr = string(payloadb)
-				req.Body = io.NopCloser(bytes.NewReader(payloadb))
-			}
-			payloadHash := sha256.Sum256([]byte(payloadStr))
-			payloadHashHex := strings.ToLower(hex.EncodeToString(payloadHash[:]))
-
-			nowUtc := time.Now().UTC()
-			timestampStr := req.Header.Get("X-CNC-Timestamp")
-			if timestampStr == "" {
-				timestampStr = fmt.Sprintf("%d", nowUtc.Unix())
-			} else {
-				timestamp, err := strconv.ParseInt(timestampStr, 10, 64)
-				if err != nil {
-					return err
-				}
-				nowUtc = time.Unix(timestamp, 0).UTC()
-			}
-
-			canonicalRequest := fmt.Sprintf("%s\n%s\n%s\n%s\n%s\n%s", method, path, queryStr, canonicalHeaders, signedHeaders, payloadHashHex)
-			canonicalRequestHash := sha256.Sum256([]byte(canonicalRequest))
-			canonicalRequestHashHex := strings.ToLower(hex.EncodeToString(canonicalRequestHash[:]))
-
-			const signAlgorithmHeader = "CNC-HMAC-SHA256"
-			stringToSign := fmt.Sprintf("%s\n%s\n%s", signAlgorithmHeader, timestampStr, canonicalRequestHashHex)
-
-			h := hmac.New(sha256.New, []byte(options.SecretKey))
-			h.Write([]byte(stringToSign))
-			signature := strings.ToLower(hex.EncodeToString(h.Sum(nil)))
-
-			req.Header.Set("Authorization", fmt.Sprintf("%s Credential=%s, SignedHeaders=%s, Signature=%s", signAlgorithmHeader, options.AccessKey, signedHeaders, signature))
-			req.Header.Set("Date", nowUtc.Format(http.TimeFormat))
-			req.Header.Set("X-CNC-Auth-Method", "AKSK")
-			req.Header.Set("X-CNC-AccessKey", options.AccessKey)
-			req.Header.Set("X-CNC-Timestamp", timestampStr)
 
 			return nil
 		})
 
-	return &Client{rc: restyClient}, nil
+	return &Client{rc: httper}, nil
 }
 
 func (c *Client) SetTimeout(timeout time.Duration) *Client {
